+++ tests/apps/courses/test_models_person.py
@@ -1,7 +1,6 @@
 """
 Unit tests for the Person model
 """
-from django.core.exceptions import ValidationError
 from django.test import TestCase
 
 from cms.api import add_plugin, create_page
@@ -17,29 +16,6 @@ class PersonModelsTestCase(TestCase):
     Unit test suite to validate the behavior of the Person model
     """
 
-    def test_models_person_fields_first_name_required(self):
-        """
-        The `first_name` field should be required
-        """
-        with self.assertRaises(ValidationError) as context:
-            PersonFactory(first_name=None)
-        self.assertEqual(context.exception.messages[0], "This field cannot be null.")
-
-    def test_models_person_fields_last_name_required(self):
-        """
-        The `last_name` field should be required
-        """
-        with self.assertRaises(ValidationError) as context:
-            PersonFactory(last_name=None)
-        self.assertEqual(context.exception.messages[0], "This field cannot be null.")
-
-    def test_models_person_fields_person_title_not_required(self):
-        """
-        The `person_title` field should not be required
-        """
-        person = PersonFactory(person_title=None)
-        self.assertIsNone(person.person_title)
-
     def test_models_person_str(self):
         """
         The string representation should be built with the page `title`
@@ -48,35 +24,9 @@ class PersonModelsTestCase(TestCase):
         page = create_page(
             "Page of Lady Louise Dupont", "courses/cms/person_detail.html", "en"
         )
-        person = PersonFactory(
-            extended_object=page,
-            first_name="Louise",
-            last_name="Dupont",
-            person_title__translation__title="Madam",
-            person_title__translation__abbreviation="Mme",
-        )
-        with self.assertNumQueries(3):
-            self.assertEqual(
-                str(person), "Person: Page of Lady Louise Dupont (Madam Louise Dupont)"
-            )
-
-    def test_models_person_get_full_name(self):
-        """
-        The get_full_name method should return title, first name and last name separated by space.
-        No SQL query should be generated.
-        """
-        person1 = PersonFactory(
-            first_name="Louise",
-            last_name="Dupont",
-            person_title__translation__title="Madam",
-            person_title__translation__abbreviation="Mme",
-        )
-        person2 = PersonFactory(
-            first_name="Jacques", last_name="Martin", person_title=None
-        )
-        with self.assertNumQueries(1):
-            self.assertEqual(person1.get_full_name(), "Madam Louise Dupont")
-            self.assertEqual(person2.get_full_name(), "Jacques Martin")
+        person = PersonFactory(extended_object=page)
+        with self.assertNumQueries(2):
+            self.assertEqual(str(person), "Person: Page of Lady Louise Dupont")
 
     def test_models_person_get_courses(self):
         """